Eye tracking
With the progress of knowledge and the improvement of the quality and accuracy of eye tracking, eye tracker systems have become one of the most important and efficient tools in the field of studying human behavior and attention. Eye tracking is a method of studying and researching where a person looks. The human brain automatically directs the eyes to the direction that attracts attention and contains valuable information. Therefore, by tracking the path of a person's eyes, it is possible to get the information that the brain is processing at that moment